+++ src/types/client/section.ts
@@ -1,3 +1,4 @@
+import type { PlaylistType } from "../api/audio/playlist";
 import type { Thumbail } from "../api/audio/thumbnail";
 import type { Breadcrumb } from "../api/catalog";
 
@@ -56,4 +57,72 @@ export type AudioSection = MinimalAudioSection & {
 export type AudioSectionList = {
   defaultSection: string;
   sections: MinimalAudioSection[];
+};
+
+export type AudioMix = {
+  /**
+   * "common" for vk mix
+   */
+  id: string;
+  description: string;
+  /**
+   * lottie background animation json url
+   */
+  lottieBgUrl?: string;
+  isTunable: boolean;
+  titles: {
+    common: string;
+    play: string;
+  };
+};
+
+export type PlaylistItemOriginal = {
+  playlistId: number;
+  ownerId: number;
+  accessKey: string;
+};
+
+export type PlaylistItemFollowed = {
+  playlistId: number;
+  ownerId: number;
+};
+
+export type PlaylistItem = {
+  id: number;
+  ownerId: number;
+  type: PlaylistType;
+  title: string;
+  description: string;
+  /**
+   * count of audios inside playlist
+   */
+  count: number;
+  followers: number;
+  plays: number;
+  /**
+   * unixtime in secods format
+   */
+  createdAt: number;
+  /**
+   * unixtime in secods format
+   */
+  updatedAt: number;
+  genres: unknown[];
+  isFollowing: boolean;
+  photo: Thumbail;
+  thumbs: Thumbail[];
+  accessKey: string;
+  original?: PlaylistItemOriginal;
+  followed?: PlaylistItemFollowed;
+  /**
+   * hex color
+   */
+  mainColor?: string;
+  subtitle?: string;
+};
+
+export type AudioSectionListWithBlocks = AudioSectionList & {
+  audioMixes: AudioMix[];
+  recentAudios: AudioItem[];
+  playlists: PlaylistItem[];
 };